+++ packages/site/web/components/Markdown/index.jsx
@@ -3,7 +3,11 @@ import { pascalCase } from 'change-case';
 import marked from 'marked';
 import Prism from 'prismjs';
 import Meta from '@/web/components/Meta';
+import { documents } from '@/site.config';
+import Codesanbox from './Codesanbox';
 import './style.scss';
+
+const isComponentPage = (page) => documents.findIndex((item) => item.key === page) === -1;
 
 export default (props) => {
   const { document, component } = props;
@@ -22,7 +26,16 @@ export default (props) => {
         Object.keys(Prism.languages).indexOf(language) > -1
           ? Prism.highlight(code, Prism.languages[language], language)
           : code;
-      return `<pre><code class="language-${language}">${highlightCode}</code></pre>`;
+
+      if (!isComponentPage(component.key)) {
+        return `<pre><code class="language-${language}">${highlightCode}</code></pre>`;
+      }
+
+      return Codesanbox({
+        code,
+        component,
+        preview: `<pre><code class="language-${language}">${highlightCode}</code></pre>`,
+      });
     };
 
     // 标题
@@ -45,17 +58,18 @@ export default (props) => {
   }, []);
 
   if (typeof document === 'string') {
+    const title = `${component.name} ${pascalCase(component.key)} - Zarm Design`;
+    const pageCls = `${component.key}-page`;
     const demoHTML = marked(document.replace(/## API\s?([^]+)/g, ''), { renderer: getRenderer() });
     const api = document.match(/## API\s?([^]+)/g);
     const apiHTML = marked(Object.prototype.toString.call(api) === '[object Array]' ? api[0] : '', {
       renderer: getRenderer(),
     });
-    const title = `${component.name} ${pascalCase(component.key)} - Zarm Design`;
 
     return (
       <>
         <Meta title={title} description={component.description || component.name} />
-        <div className={`${component.key}-page`}>
+        <div className={pageCls}>
           <div className="demo" dangerouslySetInnerHTML={{ __html: demoHTML }} />
           <div className="api" dangerouslySetInnerHTML={{ __html: apiHTML }} />
         </div>