+++ packages/site/web/pages/Design/index.jsx
@@ -1,33 +1,25 @@
-import React from 'react';
+import React, { Suspense, lazy } from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
-import Loadable from 'react-loadable';
 import { design } from '@/site.config';
 import Container from '@/web/components/Container';
-// import Footer from '@/web/components/Footer';
 import './style.scss';
-
-const LoadableComponent = (component) => {
-  return Loadable({
-    loader: component.module,
-    loading: () => null,
-  });
-};
 
 const Page = () => {
   return (
-    <Container className="design-page">
-      <main>
-        <div className="main-container markdown">
-          <Switch>
-            {design.map((doc, i) => (
-              <Route key={+i} path={`/design/${doc.key}`} component={LoadableComponent(doc)} />
-            ))}
-            <Redirect to="/" />
-          </Switch>
-        </div>
-      </main>
-      {/* <Footer /> */}
-    </Container>
+    <Suspense fallback={() => null}>
+      <Container className="design-page">
+        <main>
+          <div className="main-container markdown">
+            <Switch>
+              {design.map((doc, i) => (
+                <Route key={+i} path={`/design/${doc.key}`} component={lazy(doc.module)} />
+              ))}
+              <Redirect to="/" />
+            </Switch>
+          </div>
+        </main>
+      </Container>
+    </Suspense>
   );
 };